Resolve merge conflicts with latest base
