+++ src/goatkv/core/mod.rs
@@ -8,6 +8,4 @@ pub mod lsm_state;
 pub mod mem_table;
 pub mod sequence_number;
 pub mod skip_list;
-pub mod wal_handle;
-
 pub use kv_engine::KvEngine;
